修改bug

+++ xstart/src/main/java/com/example/start/module/controller/SysResourcesController.java
@@ -1,7 +1,9 @@
 package com.example.start.module.controller;
 
 import com.example.start.common.base.BaseController;
+import com.example.start.common.exception.ExceptionCode;
 import com.example.start.common.exception.ServiceException;
+import com.example.start.common.utils.StringUtil;
 import com.example.start.module.entity.SysResources;
 import com.example.start.module.service.SysResourcesService;
 import io.swagger.annotations.ApiImplicitParam;
@@ -30,13 +32,37 @@ public class SysResourcesController extends BaseController {
     @ApiOperation(value="新增资源", notes="根据entity对象创建资源")
     @ApiImplicitParam(name = "entity", value = "资源实体entity", required = true, dataType = "entity")
     @RequestMapping(method = RequestMethod.POST)
-    public Map<String, Object> add(@RequestBody SysResources entity) throws ServiceException {    	
+    public Map<String, Object> add(@RequestBody SysResources entity) throws ServiceException {
+        if(entity == null){
+            throw new ServiceException(ExceptionCode.ILLEGAL_PARAMETER);
+        }
+        if(StringUtil.isNull(entity.getMenuName())){
+            throw new ServiceException(ExceptionCode.ILLEGAL_PARAMETER.getCode(),"请输入菜单名！");
+        }
+        if(entity.getSort() == null){
+            throw new ServiceException(ExceptionCode.ILLEGAL_PARAMETER.getCode(),"请输入菜单排序！");
+        }
+        if(entity.getDisabled() == null){
+            throw new ServiceException(ExceptionCode.ILLEGAL_PARAMETER.getCode(),"请选择启用/禁用！");
+        }
         return success(sysResourcesService.add(entity));
     }
     @ApiOperation(value="修改资源", notes="根据entity对象编辑资源")
     @ApiImplicitParam(name = "entity", value = "资源实体entity", required = true, dataType = "entity")
     @RequestMapping(method = RequestMethod.PUT)
-    public Map<String, Object> update(@RequestBody SysResources entity) throws ServiceException {    	
+    public Map<String, Object> update(@RequestBody SysResources entity) throws ServiceException {
+        if(entity == null || entity.getId() == null){
+            throw new ServiceException(ExceptionCode.ILLEGAL_PARAMETER);
+        }
+        if(StringUtil.isNull(entity.getMenuName())){
+            throw new ServiceException(ExceptionCode.ILLEGAL_PARAMETER.getCode(),"请输入菜单名！");
+        }
+        if(entity.getSort() == null){
+            throw new ServiceException(ExceptionCode.ILLEGAL_PARAMETER.getCode(),"请输入菜单排序！");
+        }
+        if(entity.getDisabled() == null){
+            throw new ServiceException(ExceptionCode.ILLEGAL_PARAMETER.getCode(),"请选择启用/禁用！");
+        }
         return success(sysResourcesService.update(entity));
     }
 
@@ -45,7 +71,9 @@ public class SysResourcesController extends BaseController {
     @ApiImplicitParam(name = "id", value = "资源ID", required = true, dataType = "Long",paramType = "path")
     @RequestMapping(value = "{id}", method = RequestMethod.DELETE)
     public Map<String, Object> delete(@PathVariable("id") Long id) throws ServiceException {
-        
+        if(id == null){
+            throw new ServiceException(ExceptionCode.ILLEGAL_PARAMETER);
+        }
         return success(sysResourcesService.delete(id));
     }
 
@@ -56,7 +84,9 @@ public class SysResourcesController extends BaseController {
     })
     @RequestMapping(value = "{id}/{status}", method = RequestMethod.GET)
     public Map<String, Object> disabled(@PathVariable("id") Long id,@PathVariable("status") Integer status) throws ServiceException {
-
+        if(id == null || status == null){
+            throw new ServiceException(ExceptionCode.ILLEGAL_PARAMETER);
+        }
         return success(sysResourcesService.disabled(id,status));
     }
 
@@ -64,6 +94,9 @@ public class SysResourcesController extends BaseController {
     @ApiImplicitParam(name = "id", value = "资源ID", required = true, dataType = "Long",paramType = "path")
     @RequestMapping(value = "{id}", method = RequestMethod.GET)
     public Map<String, Object> get(@PathVariable("id") Long id) throws ServiceException {
+        if(id == null){
+            throw new ServiceException(ExceptionCode.ILLEGAL_PARAMETER);
+        }
         return success(sysResourcesService.findOne(id));
     }
